Struct hazelcast_rest::HazelcastRestClient

pub struct HazelcastRestClient {
    // some fields omitted
}

Hazelcast rest api client struct.

Methods

impl HazelcastRestClient

fn new(ip_address: &'static str, port: &'static str) -> HazelcastRestClient

Creates a new client struct with given address and port

fn queue_offer<T: ToString>(self: &Self, queue_name: &str, value: T) -> Result<String, Error>

Inserts an element to the named queue

fn queue_delete(self: &Self, queue_name: &str, timeout: i32) -> Result<String, Error>

Polls an element from the named queue

fn queue_size(self: &Self, queue_name: &str) -> Result<String, Error>

Gets the size of the named queue. User should unwrap and parse the resultant string to get the number.

fn map_put<T: ToString>(self: &Self, map_name: &str, key_name: &str, value: T) -> Result<String, Error>

Puts key-value to the named map. Overwrites if given key is already in map.

fn map_get(self: &Self, map_name: &str, key_name: &str) -> Result<String, Error>

Gets element with given key from given map.

fn map_remove(self: &Self, map_name: &str, key_name: &str) -> Result<String, Error>

Removes element from given map with given key

fn map_remove_all(self: &Self, map_name: &str) -> Result<String, Error>

Removes all elements of the named map.